add error library, updated endpoints to use rating

+++ back-end/src/controller/BadgeController.ts
@@ -1,8 +1,8 @@
+import * as httpError from 'http-errors';
+
 import { AppDataSource } from '../data-source';
 import { NextFunction, Request, Response } from 'express';
-import * as moment from 'moment';
 import { Badge } from '../entity/Badge';
-import { Movie } from '../entity/Movie';
 import { logger } from '../../lib/logger';
 
 export class BadgeController {
@@ -13,57 +13,46 @@ export class BadgeController {
     return this.badgeRepository.find();
   }
 
-  async one(request: Request, response: Response, next: NextFunction) {
+  async one(request: Request, response: Response, next: NextFunction): Promise<Badge> {
     const id = parseInt(request.params.id);
-
-    const badge = await this.badgeRepository.findOne({
-      where: { id }
-    });
-
-    if (!badge) {
-      return 'unregistered badge';
+    if (!id) {
+      logger.info('missing param id');
+      throw new httpError.BadRequest('missing param id');
     }
+
+    const badge = await this.badgeRepository.findOneBy({ id });
+    if (!badge) {
+      logger.info(`badge: ${id} not found`);
+      throw new httpError.NotFound(`badge: ${id} not found`);
+    }
+
     return badge;
   }
 
   async save(request: Request, response: Response, next: NextFunction) {
-    const id = parseInt(request.params.id);
-    const { rarity, icon, name, description, movieId } = request.body;
-    const movie = await AppDataSource.getRepository(Movie).findOneBy({ id: movieId });
-    let badge = await this.badgeRepository.findOneBy({ id });
-
-    const moviesAssociated = badge.movies;
-    if (!moviesAssociated.includes(movie)) {
-      moviesAssociated.push(movie);
-    }
-
-    const updatedAt = moment().format('YYYY-MM-DD HH:mm:ss');
-
-    badge = Object.assign(badge, {
-      rarity,
-      icon,
-      name,
-      description,
-      movies: moviesAssociated,
-      updated_at: updatedAt
-    });
-    logger.info('saving badge: ' + JSON.stringify(badge));
-
-    return this.badgeRepository.save(badge);
+    // rethink this
   }
 
-  async remove(request: Request, response: Response, next: NextFunction) {
+  async remove(request: Request, response: Response, next: NextFunction): Promise<void> {
     const id = parseInt(request.params.id);
+    if (!id) {
+      logger.info('missing param id');
+      throw new httpError.BadRequest('missing param id');
+    }
 
     const badgeToRemove = await this.badgeRepository.findOneBy({ id });
-
     if (!badgeToRemove) {
-      return 'this badge not exist';
+      logger.info(`badge: ${id} not found`);
+      throw new httpError.NotFound(`badge: ${id} not found`);
     }
 
-    await this.badgeRepository.remove(badgeToRemove);
+    try {
+      await this.badgeRepository.remove(badgeToRemove);
+    } catch (e) {
+      logger.info(`error removing badge: ${id}`);
+      throw new httpError.InternalServerError(`error removing badge: ${id}`);
+    }
 
-    return 'badge has been removed';
+    return;
   }
-
 }
